Move all CSS styling to their own pages
