+++ src/pages/about/styles.module.css
@@ -1,0 +1,160 @@
+/**************
+ ABOUT US PAGE 
+***************/
+
+.contentContainer {
+  max-width: 996px;
+}
+
+.aboutUsBanner {
+  padding: 80px 0 50px 0;
+  text-align: center;
+  position: relative;
+  overflow: hidden;
+  --ifm-hero-text-color: black;
+}
+
+.aboutUsBanner h1 {
+  font-family: 'Bungee', serif;
+  font-size: 32px;
+  line-height: 150%;
+  letter-spacing: 0.2em;
+  margin: 30px 0 4px 0;
+}
+.aboutUsBanner h2 {
+  font-family: 'Permanent Marker', cursive;
+  font-size: 18px;
+  line-height: 150%;
+  margin: 0 0 80px 0;
+}
+
+.aboutUsPageSection {
+  padding: 60px 0 60px 0;
+  font-size: 18px;
+  line-height: 150%;
+}
+
+.aboutUsPageSectionLightGrey {
+  background-color: #FAFAFA;
+}
+
+.aboutUsPageSection img {
+  display: block;
+  margin-left: auto;
+  margin-right: auto;
+}
+
+.aboutUsPageSection h2 {
+  text-align: center;
+  font-size: 20px;
+  margin: 30px 0;
+}
+
+.bodyLarge {
+  font-size: 18px;
+  line-height: 150%;
+  margin-bottom: 32px;
+  text-align: left;
+}
+
+@media screen and (min-width: 966px) { /* desktop */
+  .aboutUsPageWhyUs h2 {
+    text-align: left;
+  }
+}
+@media screen and (max-width: 966px) { /* mobile all */
+  .aboutUsPageWhyUs h2 {
+    margin: 50px 0;
+  }
+}
+
+.whyUsReason {
+  display: flex;
+  margin-bottom: 30px;
+}
+
+.whyUsReason>div:nth-child(1) { /* the icon column */
+  width: 140px;
+  text-align: left;
+}
+.whyUsReason>div:nth-child(1) img { /* the icon column image */
+  margin: 10px 54px 0 0;
+  width: 64px;
+}
+
+.whyUsReason>div:nth-child(2) { /* the explanation */
+  flex: 100%;
+}
+
+@media screen and (max-width: 966px) { /* mobile all */
+  .whyUsReason {
+    display: block;
+    margin-bottom: 50px;
+    text-align: center;
+  }
+  .whyUsReason>div:nth-child(1) { /* the icon column */
+    width: 100%;
+    text-align: center;
+  }
+  .whyUsReason>div:nth-child(1) img { /* the icon column image */
+    margin: 0 auto 30px auto;
+  }
+}
+
+@media screen and (min-width: 966px) { /* desktop */
+  .aboutUsPageContributors h2 {
+    text-align: left;
+  }
+}
+@media screen and (max-width: 966px) { /* mobile all */
+  .aboutUsPageContributors {
+    text-align: center;
+  }
+}
+
+.contributorCards {
+  display: flex;
+  flex-wrap: wrap;
+  margin-top: 45px;
+}
+
+.contributorCard {
+  width: calc(25% - 20px);
+  margin: 0 20px 16px 0;
+  text-align: center;
+}
+
+@media screen and (max-width: 966px) { /* mobile all */
+  .contributorCard {
+    width: calc(33.33% - 40px);
+    margin: 8px 20px;
+  }
+}
+@media screen and (max-width: 768px) {
+  .contributorCard {
+    width: calc(50% - 40px);
+    margin: 8px 20px;
+  }
+}
+@media screen and (max-width: 480px) {
+  .contributorCard {
+    width: 80%;
+    margin: 8px 9%;
+  }
+}
+
+.contributorAvatar {
+  display: block;
+}
+
+.contributorAvatarSubtitle a {
+  color: black;
+}
+
+.contributorName {
+  color: #999999;
+}
+
+.contributorFooter {
+  padding-top: 16px;
+}
