+++ src/pages/jobs/index.tsx
@@ -4,7 +4,7 @@ import React from 'react';
 import clsx from 'clsx';
 import Layout from '@theme/Layout';
 import useDocusaurusContext from '@docusaurus/useDocusaurusContext';
-import { TrackedLink } from "../trackedComponents/TrackedLink";
+import { TrackedLink } from "../../trackedComponents/TrackedLink";
 import styles from './styles.module.css';
 
 export default function Jobs() {
